Proper endpoints in integration tests Two more integration tests (negative test cases) Refactored feedback tests
